Rename pytools to analysator
This requires all scripts to be changed. A simple
"sed -i 's/pytools/analysator/g' filename.py" should to the trick.
