+++ Documentation/components/mayavi/mayavi_usage.tex
@@ -71,7 +71,7 @@ showstringspaces=false            %
 \section{Plotting the grid}
 
 \begin{python}
-import pytools as pt
+import analysator as pt
 f = pt.vlsvfile.VlsvReader('bulk.0000872.vlsv')
 grid = pt.grid.MayaviGrid(f, 'rho')
 \end{python}
