+++ Documentation/installation/installation.tex
@@ -46,7 +46,7 @@ The installation can be tested by going to the analysator folder and typing:
 \begin{verbatim}
 cd ~/analysator
 ipython
-import pytools as pt
+import analysator as pt
 \end{verbatim}
 
 The full input/output should look like this (if the installation fails, the import command should 
@@ -63,7 +63,7 @@ IPython 1.2.1 -- An enhanced Interactive Python.
 help      -> Python's own help system.
 object?   -> Details about 'object', use 'object??' for extra details.
 
-In [1]: import pytools as pt
+In [1]: import analysator as pt
 
 In [2]:
