+++ Documentation/particle_pusher/particlepusher.tex
@@ -34,7 +34,7 @@ Starting up the particle pusher (Use VlasiatorReader, not VlsvReader):
 \begin{verbatim}
 ipython
 
-In [2]: import pytools as pt
+In [2]: import analysator as pt
 
 In [3]: f = pt.vlsvfile.VlasiatorReader('bulk.0001480.vlsv')
