+++ Documentation/presentations/analysator.tex
@@ -111,7 +111,7 @@ showstringspaces=false            %
  \begin{python}[basicstyle=\tiny]
   # We are interested in reading data from the cell whose ID is 75
  
-  import pytools as pt # Import Analysator
+  import analysator as pt # Import Analysator
   
   # Open a vlsv file for reading
   vlsvReader = pt.vlsvfile.VlsvReader('bulk.0003710.vlsv')
@@ -163,7 +163,7 @@ showstringspaces=false            %
  
  \begin{python}[basicstyle=\tiny]
   # Import analysator
-  import pytools as pt
+  import analysator as pt
   
   vlsvReader = pt.vlsvfile.VlsvReader('test.vlsv') #What?
   
@@ -189,7 +189,7 @@ showstringspaces=false            %
  \frametitle{Navigation (the \emph{most} important part)}
  \begin{python}[basicstyle=\tiny]
   # Import analysator
-  import pytools as pt
+  import analysator as pt
   
   pt.
  \end{python}
@@ -205,7 +205,7 @@ pt.filemanagement  pt.miscellaneous   pt.vlsvfile
  \frametitle{Navigation (the \emph{most} important part)}
  \begin{python}[basicstyle=\tiny]
   # Import analysator
-  import pytools as pt
+  import analysator as pt
   
   pt.calculations.
  \end{python}
@@ -227,7 +227,7 @@ pt.calculations.VariableInfo
  \frametitle{Navigation (the \emph{most} important part)}
  \begin{python}[basicstyle=\tiny]
   # Import analysator
-  import pytools as pt
+  import analysator as pt
   
   pt.calculations.cut_through?
  \end{python}
@@ -287,7 +287,7 @@ format: [cell ids, distances, coordinates]
  \frametitle{Basic usage}
  \begin{python}[basicstyle=\tiny]
   # Import analysator
-  import pytools as pt
+  import analysator as pt
   
   # Open a vlsv file
   vlsvReader = pt.vlsvfile.VlsvReader('test.vlsv')
@@ -384,7 +384,7 @@ format: [cell ids, distances, coordinates]
  Plotting can be done simply by feeding a vlsv file to the grid class as follows:
  \begin{python}[basicstyle=\tiny]
   # Import Analysator
-  import pytools as pt
+  import analysator as pt
   
   # Open a vlsv file
   vlsvReader = pt.vlsvfile.VlsvReader('test.vlsv')
@@ -532,7 +532,7 @@ format: [cell ids, distances, coordinates]
  \frametitle{Example from \emph{calculations}}
  \begin{python}[basicstyle=\tiny]
   # Import analysator
-  import pytools as pt
+  import analysator as pt
   
   # Open a vlsv file
   vlsvReader = pt.vlsvfile.vlsvReader('testfile.vlsv')
