+++ analysator.py
