+++ examples/multi_panel_plot.py
@@ -1,4 +1,4 @@
-import pytools as pt
+import analysator as pt
 import numpy as np
 import matplotlib.pyplot as plt
 import colormaps as cmaps
