+++ miscellaneous/slams.py
@@ -22,7 +22,7 @@
 # 
 
 import numpy as np
-import pytools as pt
+import analysator as pt
 import os
 import copy
 import matplotlib.pyplot as plt
