+++ pyCalculations/intpol_points.py
@@ -35,7 +35,7 @@ def vlsv_intpol_points(vlsvReader,points,varlist,operator="pass",interpolation_o
        :returns: A tuple with output: (coordinates,variable_values,header_string)
        .. code-block:: python
           # Example:
-          import pytools as pt
+          import analysator as pt
           import numpy as np
           f=pt.vlsvfile.VlsvReader(file_name="state00040000.vlsv")
           mesh_limits = f.get_spatial_mesh_extent()
