+++ pyPlots/plot_helpers.py
@@ -21,7 +21,7 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 # 
 
-import pytools as pt
+import analysator as pt
 import numpy as np
 from rotation import rotateTensorToVector
