+++ pyPlots/plot_isosurface.py
@@ -22,7 +22,7 @@
 # 
 
 import matplotlib
-import pytools as pt
+import analysator as pt
 import numpy as np
 import matplotlib.pyplot as plt
