+++ pyPlots/plot_threeslice.py
@@ -1,5 +1,5 @@
 import matplotlib
-import pytools as pt
+import analysator as pt
 import numpy as np
 import matplotlib.pyplot as plt
 import scipy
@@ -158,7 +158,7 @@ def plot_threeslice(filename=None,
     :kword step:        output step index, used for constructing output (and possibly input) filename                    
 
     :kword var:         variable to plot, e.g. rho, RhoBackstream, beta, Temperature, MA, Mms, va, vms,
-                        E, B, v, V or others. Accepts any variable known by analysator/pytools.
+                        E, B, v, V or others. Accepts any variable known by analysator/analysator.
                         Per-population variables are simply given as "proton/rho" etc
     :kword operator:    Operator to apply to variable: None, x, y, or z. Vector variables return either
                         the queried component, or otherwise the magnitude.
