+++ scripts/plot_jet_criteria.py
@@ -21,7 +21,7 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 # 
 
-import pytools as pt
+import analysator as pt
 import sys, os, socket
 import numpy as np
