+++ testpackage/testpackage_vdf.py
@@ -1,4 +1,4 @@
-import pytools as pt
+import analysator as pt
 import sys, os
 import numpy as np
 import traceback
