+++ tools/vlsvintpol.py
@@ -22,7 +22,7 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 # 
 
-import pytools as pt
+import analysator as pt
 import numpy as np
 import sys
 import argparse
