+++ trash_can/plot_veldist.py
@@ -35,7 +35,7 @@ import matplotlib
 matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import matplotlib.colors as colors
-import pytools as pt
+import analysator as pt
 import numpy as np
 import operator as oper
 from rotation import *
